requirements.txt
snowflake==0.6.0
tabulate==0.9.0
python-dotenv==1.0.0
loguru==0.6.0
terminaltables==3.1.0
